team-section added (todo : connect to appwrite)
